backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
kafka-python
